% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co_8hr_avg}
\alias{co_8hr_avg}
\title{co_8hr_avg}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44682 rows and 3 columns.
}
\usage{
co_8hr_avg
}
\description{
Carbon Monoxide 8-hour Average Concentrations
}
\keyword{datasets}
