% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co_dvs_anc}
\alias{co_dvs_anc}
\title{co_dvs_anc}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 122 rows and 4 columns.
}
\usage{
co_dvs_anc
}
\description{
Carbon Monoxide Design Values for Anchorage
}
\keyword{datasets}
