% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co_dvs_fbx}
\alias{co_dvs_fbx}
\title{co_dvs_fbx}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 96 rows and 4 columns.
}
\usage{
co_dvs_fbx
}
\description{
Carbon Monoxide Design Values for Fairbanks
}
\keyword{datasets}
