% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naaqs_data}
\alias{naaqs_data}
\title{naaqs data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 5 columns.
}
\usage{
naaqs_data
}
\description{
NAAQS PM2.5 Table Values from 1997 to Present
}
\keyword{datasets}
