% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncore_no2_1hr_max}
\alias{ncore_no2_1hr_max}
\title{ncore_no2_1hr_max}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1919 rows and 3 columns.
}
\usage{
ncore_no2_1hr_max
}
\description{
Nitrogen Dioxide 1-hour Maximum Concentrations
}
\keyword{datasets}
