% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncore_o3_8hr_avg}
\alias{ncore_o3_8hr_avg}
\title{ncore_o3_8hr_avg}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4530 rows and 3 columns.
}
\usage{
ncore_o3_8hr_avg
}
\description{
Ozone 8-hour Average Concentrations
}
\keyword{datasets}
