% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{no2_dvs}
\alias{no2_dvs}
\title{no2_dvs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 4 columns.
}
\usage{
no2_dvs
}
\description{
Nitrogen Dioxide Design Values
}
\keyword{datasets}
