% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{o3_dvs_anc}
\alias{o3_dvs_anc}
\title{o3_dvs_anc}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 4 columns.
}
\usage{
o3_dvs_anc
}
\description{
Ozone Design Values for Anchorage
}
\keyword{datasets}
