% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{o3_dvs_fbx}
\alias{o3_dvs_fbx}
\title{o3_dvs_fbx}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26 rows and 4 columns.
}
\usage{
o3_dvs_fbx
}
\description{
Ozone Design Values for Fairbanks
}
\keyword{datasets}
