% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{o3_dvs_ms}
\alias{o3_dvs_ms}
\title{o3_dvs_ms}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 4 columns.
}
\usage{
o3_dvs_ms
}
\description{
Ozone Design Values for Mat-Su
}
\keyword{datasets}
