% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm10_dvs_anc}
\alias{pm10_dvs_anc}
\title{pm10_dvs_anc}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 172 rows and 4 columns.
}
\usage{
pm10_dvs_anc
}
\description{
PM10 Design Values for Anchorage
}
\keyword{datasets}
