% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm10_dvs_ms}
\alias{pm10_dvs_ms}
\title{pm10_dvs_ms}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 78 rows and 4 columns.
}
\usage{
pm10_dvs_ms
}
\description{
PM10 Design Values for Mat-Su
}
\keyword{datasets}
