% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm10_dvs_sw}
\alias{pm10_dvs_sw}
\title{pm10_dvs_sw}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 4 columns.
}
\usage{
pm10_dvs_sw
}
\description{
PM10 Design Values for Bethel
}
\keyword{datasets}
