% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm10_import_named}
\alias{pm10_import_named}
\title{pm10_import_named}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55132 rows and 3 columns.
}
\usage{
pm10_import_named
}
\description{
PM10 24-hour Average Concentrations
}
\keyword{datasets}
