% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm25_dvs_ms}
\alias{pm25_dvs_ms}
\title{pm25_dvs_ms}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 156 rows and 4 columns.
}
\usage{
pm25_dvs_ms
}
\description{
PM2.5 Design Values for Mat-Su
}
\keyword{datasets}
