% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm25_dvs_np}
\alias{pm25_dvs_np}
\title{pm25_dvs_np}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 4 columns.
}
\usage{
pm25_dvs_np
}
\description{
PM2.5 Design Values for North Pole
}
\keyword{datasets}
