% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm25_dvs_sw}
\alias{pm25_dvs_sw}
\title{pm25_dvs_sw}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 4 columns.
}
\usage{
pm25_dvs_sw
}
\description{
PM2.5 Design Values for Bethel
}
\keyword{datasets}
