% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm25_import_named}
\alias{pm25_import_named}
\title{pm25_import_named}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58292 rows and 3 columns.
}
\usage{
pm25_import_named
}
\description{
PM2.5 24-hour Average Concentrations
}
\keyword{datasets}
