% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{so2_1hr_max}
\alias{so2_1hr_max}
\title{so2_1hr_max}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5930 rows and 3 columns.
}
\usage{
so2_1hr_max
}
\description{
Sulfur Dioxide 1-hour Maximum Concentrations
}
\keyword{datasets}
