% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{so2_dvs}
\alias{so2_dvs}
\title{so2_dvs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 28 rows and 4 columns.
}
\usage{
so2_dvs
}
\description{
Sulfur Dioxide Design Values
}
\keyword{datasets}
